+++ contrib/mesos/pkg/scheduler/framework_test.go
@@ -26,11 +26,12 @@ import (
 	"k8s.io/kubernetes/contrib/mesos/pkg/offers"
 	"k8s.io/kubernetes/contrib/mesos/pkg/proc"
 	schedcfg "k8s.io/kubernetes/contrib/mesos/pkg/scheduler/config"
-	"k8s.io/kubernetes/contrib/mesos/pkg/scheduler/podtask"
 	"k8s.io/kubernetes/contrib/mesos/pkg/scheduler/slave"
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/client/cache"
 	"k8s.io/kubernetes/contrib/mesos/pkg/scheduler/mock"
+	"k8s.io/kubernetes/contrib/mesos/pkg/scheduler/types"
+	"k8s.io/kubernetes/contrib/mesos/pkg/scheduler/podtask"
 )
 
 //get number of non-expired offers from  offer registry
@@ -82,12 +83,19 @@ func (r *mockRegistrator) Register(hostName string, labels map[string]string) (b
 	}
 }
 
+func mockScheduler() types.Scheduler {
+	mockScheduler := &types.MockScheduler{}
+	reg := podtask.NewInMemoryRegistry()
+	mockScheduler.On("Tasks").Return(reg)
+	return mockScheduler
+}
+
 //test adding of ressource offer, should be added to offer registry and slaves
 func TestResourceOffer_Add(t *testing.T) {
 	assert := assert.New(t)
 
 	registrator := &mockRegistrator{cache.NewStore(cache.MetaNamespaceKeyFunc)}
-	testScheduler := &MesosScheduler{
+	testFramework := &Framework{
 		offers: offers.CreateRegistry(offers.RegistryConfig{
 			Compat: func(o *mesos.Offer) bool {
 				return true
@@ -102,37 +110,38 @@ func TestResourceOffer_Add(t *testing.T) {
 		}),
 		slaveHostNames:  slave.NewRegistry(),
 		nodeRegistrator: registrator,
+		sched:       mockScheduler(),
 	}
 
 	hostname := "h1"
 	offerID1 := util.NewOfferID("test1")
 	offer1 := &mesos.Offer{Id: offerID1, Hostname: &hostname, SlaveId: util.NewSlaveID(hostname)}
 	offers1 := []*mesos.Offer{offer1}
-	testScheduler.ResourceOffers(nil, offers1)
+	testFramework.ResourceOffers(nil, offers1)
 	assert.Equal(1, len(registrator.store.List()))
 
-	assert.Equal(1, getNumberOffers(testScheduler.offers))
+	assert.Equal(1, getNumberOffers(testFramework.offers))
 	//check slave hostname
-	assert.Equal(1, len(testScheduler.slaveHostNames.SlaveIDs()))
+	assert.Equal(1, len(testFramework.slaveHostNames.SlaveIDs()))
 
 	//add another offer
 	hostname2 := "h2"
 	offer2 := &mesos.Offer{Id: util.NewOfferID("test2"), Hostname: &hostname2, SlaveId: util.NewSlaveID(hostname2)}
 	offers2 := []*mesos.Offer{offer2}
-	testScheduler.ResourceOffers(nil, offers2)
+	testFramework.ResourceOffers(nil, offers2)
 
 	//check it is stored in registry
-	assert.Equal(2, getNumberOffers(testScheduler.offers))
+	assert.Equal(2, getNumberOffers(testFramework.offers))
 
 	//check slave hostnames
-	assert.Equal(2, len(testScheduler.slaveHostNames.SlaveIDs()))
+	assert.Equal(2, len(testFramework.slaveHostNames.SlaveIDs()))
 }
 
 //test adding of ressource offer, should be added to offer registry and slavesf
 func TestResourceOffer_Add_Rescind(t *testing.T) {
 	assert := assert.New(t)
 
-	testScheduler := &MesosScheduler{
+	testFramework := &Framework{
 		offers: offers.CreateRegistry(offers.RegistryConfig{
 			Compat: func(o *mesos.Offer) bool {
 				return true
@@ -146,41 +155,42 @@ func TestResourceOffer_Add_Rescind(t *testing.T) {
 			ListenerDelay: schedcfg.DefaultListenerDelay,
 		}),
 		slaveHostNames: slave.NewRegistry(),
+		sched:       mockScheduler(),
 	}
 
 	hostname := "h1"
 	offerID1 := util.NewOfferID("test1")
 	offer1 := &mesos.Offer{Id: offerID1, Hostname: &hostname, SlaveId: util.NewSlaveID(hostname)}
 	offers1 := []*mesos.Offer{offer1}
-	testScheduler.ResourceOffers(nil, offers1)
+	testFramework.ResourceOffers(nil, offers1)
 
-	assert.Equal(1, getNumberOffers(testScheduler.offers))
+	assert.Equal(1, getNumberOffers(testFramework.offers))
 
 	//check slave hostname
-	assert.Equal(1, len(testScheduler.slaveHostNames.SlaveIDs()))
+	assert.Equal(1, len(testFramework.slaveHostNames.SlaveIDs()))
 
 	//add another offer
 	hostname2 := "h2"
 	offer2 := &mesos.Offer{Id: util.NewOfferID("test2"), Hostname: &hostname2, SlaveId: util.NewSlaveID(hostname2)}
 	offers2 := []*mesos.Offer{offer2}
-	testScheduler.ResourceOffers(nil, offers2)
+	testFramework.ResourceOffers(nil, offers2)
 
-	assert.Equal(2, getNumberOffers(testScheduler.offers))
+	assert.Equal(2, getNumberOffers(testFramework.offers))
 
 	//check slave hostnames
-	assert.Equal(2, len(testScheduler.slaveHostNames.SlaveIDs()))
+	assert.Equal(2, len(testFramework.slaveHostNames.SlaveIDs()))
 
 	//next whether offers can be rescinded
-	testScheduler.OfferRescinded(nil, offerID1)
-	assert.Equal(1, getNumberOffers(testScheduler.offers))
+	testFramework.OfferRescinded(nil, offerID1)
+	assert.Equal(1, getNumberOffers(testFramework.offers))
 
 	//next whether offers can be rescinded
-	testScheduler.OfferRescinded(nil, util.NewOfferID("test2"))
+	testFramework.OfferRescinded(nil, util.NewOfferID("test2"))
 	//walk offers again and check it is removed from registry
-	assert.Equal(0, getNumberOffers(testScheduler.offers))
+	assert.Equal(0, getNumberOffers(testFramework.offers))
 
 	//remove non existing ID
-	testScheduler.OfferRescinded(nil, util.NewOfferID("notExist"))
+	testFramework.OfferRescinded(nil, util.NewOfferID("notExist"))
 }
 
 //test that when a slave is lost we remove all offers
@@ -188,7 +198,7 @@ func TestSlave_Lost(t *testing.T) {
 	assert := assert.New(t)
 
 	//
-	testScheduler := &MesosScheduler{
+	testFramework := &Framework{
 		offers: offers.CreateRegistry(offers.RegistryConfig{
 			Compat: func(o *mesos.Offer) bool {
 				return true
@@ -199,44 +209,45 @@ func TestSlave_Lost(t *testing.T) {
 			ListenerDelay: schedcfg.DefaultListenerDelay,
 		}),
 		slaveHostNames: slave.NewRegistry(),
+		sched:       mockScheduler(),
 	}
 
 	hostname := "h1"
 	offer1 := &mesos.Offer{Id: util.NewOfferID("test1"), Hostname: &hostname, SlaveId: util.NewSlaveID(hostname)}
 	offers1 := []*mesos.Offer{offer1}
-	testScheduler.ResourceOffers(nil, offers1)
+	testFramework.ResourceOffers(nil, offers1)
 	offer2 := &mesos.Offer{Id: util.NewOfferID("test2"), Hostname: &hostname, SlaveId: util.NewSlaveID(hostname)}
 	offers2 := []*mesos.Offer{offer2}
-	testScheduler.ResourceOffers(nil, offers2)
+	testFramework.ResourceOffers(nil, offers2)
 
 	//add another offer from different slaveID
 	hostname2 := "h2"
 	offer3 := &mesos.Offer{Id: util.NewOfferID("test3"), Hostname: &hostname2, SlaveId: util.NewSlaveID(hostname2)}
 	offers3 := []*mesos.Offer{offer3}
-	testScheduler.ResourceOffers(nil, offers3)
+	testFramework.ResourceOffers(nil, offers3)
 
 	//test precondition
-	assert.Equal(3, getNumberOffers(testScheduler.offers))
-	assert.Equal(2, len(testScheduler.slaveHostNames.SlaveIDs()))
+	assert.Equal(3, getNumberOffers(testFramework.offers))
+	assert.Equal(2, len(testFramework.slaveHostNames.SlaveIDs()))
 
 	//remove first slave
-	testScheduler.SlaveLost(nil, util.NewSlaveID(hostname))
+	testFramework.SlaveLost(nil, util.NewSlaveID(hostname))
 
 	//offers should be removed
-	assert.Equal(1, getNumberOffers(testScheduler.offers))
+	assert.Equal(1, getNumberOffers(testFramework.offers))
 	//slave hostnames should still be all present
-	assert.Equal(2, len(testScheduler.slaveHostNames.SlaveIDs()))
+	assert.Equal(2, len(testFramework.slaveHostNames.SlaveIDs()))
 
 	//remove second slave
-	testScheduler.SlaveLost(nil, util.NewSlaveID(hostname2))
+	testFramework.SlaveLost(nil, util.NewSlaveID(hostname2))
 
 	//offers should be removed
-	assert.Equal(0, getNumberOffers(testScheduler.offers))
+	assert.Equal(0, getNumberOffers(testFramework.offers))
 	//slave hostnames should still be all present
-	assert.Equal(2, len(testScheduler.slaveHostNames.SlaveIDs()))
+	assert.Equal(2, len(testFramework.slaveHostNames.SlaveIDs()))
 
 	//try to remove non existing slave
-	testScheduler.SlaveLost(nil, util.NewSlaveID("notExist"))
+	testFramework.SlaveLost(nil, util.NewSlaveID("notExist"))
 
 }
 
@@ -245,7 +256,7 @@ func TestDisconnect(t *testing.T) {
 	assert := assert.New(t)
 
 	//
-	testScheduler := &MesosScheduler{
+	testFramework := &Framework{
 		offers: offers.CreateRegistry(offers.RegistryConfig{
 			Compat: func(o *mesos.Offer) bool {
 				return true
@@ -256,29 +267,30 @@ func TestDisconnect(t *testing.T) {
 			ListenerDelay: schedcfg.DefaultListenerDelay,
 		}),
 		slaveHostNames: slave.NewRegistry(),
+		sched:       mockScheduler(),
 	}
 
 	hostname := "h1"
 	offer1 := &mesos.Offer{Id: util.NewOfferID("test1"), Hostname: &hostname, SlaveId: util.NewSlaveID(hostname)}
 	offers1 := []*mesos.Offer{offer1}
-	testScheduler.ResourceOffers(nil, offers1)
+	testFramework.ResourceOffers(nil, offers1)
 	offer2 := &mesos.Offer{Id: util.NewOfferID("test2"), Hostname: &hostname, SlaveId: util.NewSlaveID(hostname)}
 	offers2 := []*mesos.Offer{offer2}
-	testScheduler.ResourceOffers(nil, offers2)
+	testFramework.ResourceOffers(nil, offers2)
 
 	//add another offer from different slaveID
 	hostname2 := "h2"
 	offer3 := &mesos.Offer{Id: util.NewOfferID("test2"), Hostname: &hostname2, SlaveId: util.NewSlaveID(hostname2)}
 	offers3 := []*mesos.Offer{offer3}
-	testScheduler.ResourceOffers(nil, offers3)
+	testFramework.ResourceOffers(nil, offers3)
 
 	//disconnect
-	testScheduler.Disconnected(nil)
+	testFramework.Disconnected(nil)
 
 	//all offers should be removed
-	assert.Equal(0, getNumberOffers(testScheduler.offers))
+	assert.Equal(0, getNumberOffers(testFramework.offers))
 	//slave hostnames should still be all present
-	assert.Equal(2, len(testScheduler.slaveHostNames.SlaveIDs()))
+	assert.Equal(2, len(testFramework.slaveHostNames.SlaveIDs()))
 }
 
 //test we can handle different status updates, TODO check state transitions
@@ -288,7 +300,7 @@ func TestStatus_Update(t *testing.T) {
 	// setup expectations
 	mockdriver.On("KillTask", util.NewTaskID("test-task-001")).Return(mesos.Status_DRIVER_RUNNING, nil)
 
-	testScheduler := &MesosScheduler{
+	testFramework := &Framework{
 		offers: offers.CreateRegistry(offers.RegistryConfig{
 			Compat: func(o *mesos.Offer) bool {
 				return true
@@ -300,26 +312,26 @@ func TestStatus_Update(t *testing.T) {
 		}),
 		slaveHostNames: slave.NewRegistry(),
 		driver:         &mockdriver,
-		taskRegistry:   podtask.NewInMemoryRegistry(),
+		sched:       mockScheduler(),
 	}
 
 	taskStatus_task_starting := util.NewTaskStatus(
 		util.NewTaskID("test-task-001"),
 		mesos.TaskState_TASK_RUNNING,
 	)
-	testScheduler.StatusUpdate(testScheduler.driver, taskStatus_task_starting)
+	testFramework.StatusUpdate(testFramework.driver, taskStatus_task_starting)
 
 	taskStatus_task_running := util.NewTaskStatus(
 		util.NewTaskID("test-task-001"),
 		mesos.TaskState_TASK_RUNNING,
 	)
-	testScheduler.StatusUpdate(testScheduler.driver, taskStatus_task_running)
+	testFramework.StatusUpdate(testFramework.driver, taskStatus_task_running)
 
 	taskStatus_task_failed := util.NewTaskStatus(
 		util.NewTaskID("test-task-001"),
 		mesos.TaskState_TASK_FAILED,
 	)
-	testScheduler.StatusUpdate(testScheduler.driver, taskStatus_task_failed)
+	testFramework.StatusUpdate(testFramework.driver, taskStatus_task_failed)
 
 	//assert that mock was invoked
 	mockdriver.AssertExpectations(t)
