+++ contrib/mesos/pkg/scheduler/operations/schedulerloop.go
@@ -17,19 +17,11 @@ limitations under the License.
 package operations
 
 import (
-	"net/http"
 	"time"
 
 	log "github.com/golang/glog"
-	"k8s.io/kubernetes/contrib/mesos/pkg/backoff"
-	"k8s.io/kubernetes/contrib/mesos/pkg/queue"
 	"k8s.io/kubernetes/contrib/mesos/pkg/runtime"
-	"k8s.io/kubernetes/contrib/mesos/pkg/scheduler/config"
-	"k8s.io/kubernetes/contrib/mesos/pkg/scheduler/podtask"
-	"k8s.io/kubernetes/contrib/mesos/pkg/scheduler/queuer"
-	types "k8s.io/kubernetes/contrib/mesos/pkg/scheduler/types"
 	"k8s.io/kubernetes/pkg/api"
-	"k8s.io/kubernetes/pkg/client/cache"
 	"k8s.io/kubernetes/pkg/client/record"
 	client "k8s.io/kubernetes/pkg/client/unversioned"
 )
@@ -53,47 +45,6 @@ type SchedulerLoop struct {
 	recorder  record.EventRecorder
 	client    *client.Client
 	started   chan<- struct{} // startup latch
-}
-
-func NewScheduler(c *config.Config, fw types.Framework, client *client.Client, recorder record.EventRecorder,
-	terminate <-chan struct{}, mux *http.ServeMux, podsWatcher *cache.ListWatch) (SchedulerLoopInterface, *PodReconciler) {
-
-	// Watch and queue pods that need scheduling.
-	updates := make(chan queue.Entry, c.UpdatesBacklog)
-	podUpdates := &podStoreAdapter{queue.NewHistorical(updates)}
-	reflector := cache.NewReflector(podsWatcher, &api.Pod{}, podUpdates, 0)
-
-	// lock that guards critial sections that involve transferring pods from
-	// the store (cache) to the scheduling queue; its purpose is to maintain
-	// an ordering (vs interleaving) of operations that's easier to reason about.
-
-	q := queuer.New(podUpdates)
-
-	algorithm := NewSchedulerAlgorithm(fw, podUpdates)
-
-	podDeleter := NewDeleter(fw, q)
-
-	podReconciler := NewPodReconciler(fw, client, q, podDeleter)
-
-	bo := backoff.New(c.InitialPodBackoff.Duration, c.MaxPodBackoff.Duration)
-	errorHandler := NewErrorHandler(fw, bo, q)
-
-	binder := NewBinder(fw)
-
-	startLatch := make(chan struct{})
-	eventBroadcaster := record.NewBroadcaster()
-
-	runtime.On(startLatch, func() {
-		eventBroadcaster.StartRecordingToSink(client.Events(""))
-		reflector.Run() // TODO(jdef) should listen for termination
-		podDeleter.Run(updates, terminate)
-		q.Run(terminate)
-
-		q.InstallDebugHandlers(mux)
-		podtask.InstallDebugHandlers(fw.Tasks(), mux)
-	})
-
-	return NewSchedulerLoop(client, algorithm, recorder, q.Yield, errorHandler.Error, binder, startLatch), podReconciler
 }
 
 func NewSchedulerLoop(client *client.Client, algorithm *SchedulerAlgorithm,
