+++ contrib/mesos/pkg/scheduler/podstoreadapter.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package operations
+package scheduler
 
 import (
 	"k8s.io/kubernetes/contrib/mesos/pkg/queue"
